+++ test/functional-tests/node-transformers/converting-transformers/object-expression-keys-transformer/fixtures/return-statement-sequence-expression-side-effect.js
@@ -1,0 +1,9 @@
+function compErr(ys) {
+    let min = Infinity;
+
+    function aux(y) {
+        if (y < min) min = y;
+    }
+
+    return (aux(ys), { min });
+}